setvar modemsetup = ATE0... overrides ACTS driver default.
Preserve last timecode in ACTS driver (ntpq -ccv).Tolerate previous
ATE1 state when sending ACTS setup.Enable raw tty line discipline in
Windows port.

Allow tty open/close/open to succeed on Windows port.Enable ACTS and
CHU reference clock drivers on Windows.

+++ libntp/dolfptoa.c
@@ -60,7 +60,7 @@ dolfptoa(
 		dec   += 3;
 		cpdec += 3;
 	}
-	if (dec > sizeof(cbuf) - (cpend - cbuf))
+	if ((size_t)dec > sizeof(cbuf) - (cpend - cbuf))
 		dec = sizeof(cbuf) - (cpend - cbuf);
 	
 	/*
@@ -120,10 +120,10 @@ dolfptoa(
 
 	bp = buf;
 	if (neg)
-	    *bp++ = '-';
+		*bp++ = '-';
 	while (cp < cpend) {
 		if (cp == cpdec)
-		    *bp++ = '.';
+			*bp++ = '.';
 		*bp++ = (char)(*cp++) + '0';
 	}
 	*bp = '\0';
